temp/1/requirements.txt
python-dotenv==1.0.1
pandas==2.2.1
numpy==1.26.4
requests==2.31.0
websockets==15.0.1
pybit==5.9.0
tabulate==0.9.0
pytest==8.0.2
requests-mock==1.11.0
pycryptodome==3.21.0
aiologger==0.7.0
psutil==5.9.8
ta==0.10.2
